refactor: fly modes
